backend/requirements-raw.txt
aiohttp==3.9.5
aiosignal==1.4.0
alembic==1.16.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cmdstanpy==1.2.5
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
fastapi==0.115.6
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
holidays==0.79
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
multidict==6.6.4
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.2.3
parso==0.8.5
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
prophet==1.1.5
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
requests==2.32.3
rsa==4.9.1
sec-api==1.0.32
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.36
stack-data==0.6.3
stanio==0.5.1
starlette==0.41.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.65
